updated IN endpoint from 1 -> 2 bytes, updated led.c to use the 9th bit for the status LED, tweaked python scripts to detect larger IN read

+++ CompositeImplementation/src/modules/led.h
@@ -45,7 +45,7 @@ void led_statusOn     (void);
 void led_statusOff    (void);
 void led_statusToggle (void);
 
-uint8_t led_getMap(void);
+uint16_t led_getMap(void);
 // bool led_getStateIndex(uint8_t index);
 // bool led_getStateMask(uint8_t mask);
